+++ src/pages/seller/dashboard/components/DashboardHeader.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const DashboardHeader = ({ title }) => {
+  return (
+    <h1 className="text-2xl font-semibold mb-6 pb-2 border-b-2">{title}</h1>
+  );
+};
+
+export default DashboardHeader;
