+++ src/pages/seller/dashboard/utils/ChartRegistry.jsx
@@ -1,0 +1,12 @@
+import { Chart as ChartJS, CategoryScale, LinearScale, BarElement, Title, Tooltip, Legend, ArcElement } from "chart.js";
+
+// 차트 라이브러리 등록
+ChartJS.register(
+  CategoryScale,
+  LinearScale, 
+  BarElement, 
+  Title, 
+  Tooltip, 
+  Legend, 
+  ArcElement
+);
